test: replace ava-spec by vanilla ava :rocket:
