+++ package.json
@@ -36,7 +36,6 @@
   "devDependencies": {
     "@-k/eslint-plugin": "^0.3.1",
     "ava": "^2.4.0",
-    "ava-spec": "^1.1.1",
     "codecov": "^3.6.5",
     "eslint": "^6.8.0",
     "nyc": "^15.0.0",
